AI follow the Delfland example
