+++ data/AmstelGooienVecht/modellen/.gitignore
@@ -1,1 +1,3 @@
 /AmstelGooienVecht_parameterized
+/AmstelGooienVecht_feedback
+/AmstelGooienVecht_profiles
