+++ data/HollandsNoorderkwartier/modellen/.gitignore
@@ -1,1 +1,3 @@
 /HollandsNoorderkwartier_parameterized
+/HollandsNoorderkwartier_feedback
+/HollandsNoorderkwartier_profiles
