+++ data/HollandseDelta/modellen/.gitignore
@@ -1,1 +1,3 @@
 /HollandseDelta_parameterized
+/HollandseDelta_feedback
+/HollandseDelta_profiles
