+++ data/Rijnland/modellen/.gitignore
@@ -1,1 +1,3 @@
 /Rijnland_parameterized
+/Rijnland_feedback
+/Rijnland_profiles
