+++ data/Rivierenland/modellen/.gitignore
@@ -1,1 +1,3 @@
 /Rivierenland_parameterized
+/Rivierenland_feedback
+/Rivierenland_profiles
