+++ data/Scheldestromen/modellen/.gitignore
@@ -1,1 +1,3 @@
 /Scheldestromen_parameterized
+/Scheldestromen_feedback
+/Scheldestromen_profiles
