+++ data/SchielandendeKrimpenerwaard/modellen/.gitignore
@@ -1,1 +1,3 @@
 /SchielandendeKrimpenerwaard_parameterized
+/SchielandendeKrimpenerwaard_feedback
+/SchielandendeKrimpenerwaard_profiles
