+++ data/WetterskipFryslan/modellen/.gitignore
@@ -1,1 +1,3 @@
 /WetterskipFryslan_parameterized
+/WetterskipFryslan_feedback
+/WetterskipFryslan_profiles
