+++ data/Zuiderzeeland/modellen/.gitignore
@@ -1,1 +1,3 @@
 /Zuiderzeeland_parameterized
+/Zuiderzeeland_feedback
+/Zuiderzeeland_profiles
